HTML, CSS and Flask Codes
The index.html file contains the HTML and CSS code for the front end of the dashboard. The app.py file contains the code acting upon the routed model and form designed in the HTML file. The quality.py file contains the back-end code to predict the quality class of the vehicular engine.